added payment

+++ app/src/main/java/com/nihatmahammadli/abbmobile/domain/model/UiCard.kt
@@ -4,5 +4,5 @@ data class UiCard(
     val cardNumber: String,
     val expiryDate: String,
     val cvv: String,
-    val balance: Int = 0
+    val balance: Double  = 0.0
 )
